pyproject.toml
[project]
name = "pattern-recognition-b"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "accelerate>=1.8.1",
    "datasets>=3.6.0",
    "peft>=0.15.2",
    "pillow>=11.3.0",
    "qwen-vl-utils>=0.0.11",
    "setuptools>=80.9.0",
    "torchvision>=0.22.1",
    "tqdm>=4.67.1",
    "transformers",
]

[tool.uv.sources]
transformers = { git = "https://github.com/huggingface/transformers" }
